Lazy eval CAPTURE_OPTIONS
As-is, the code is a bit unhappy that when we source the config, we can't properly evaluate CAPTURE_OPTIONS due to ss_flag not being defined. This moves us to using lazy evaluation. It also adjusts snap.sh to error out when there are issues. Finally, it makes snap.sh better at picking up config changes.